requirements.txt
discord.py[voice]
plexapi
python-dotenv
aiohttp
pynacl